Environmental sensor OIC resource (#48)
* [ocf-servers] Add an environmental sensor OIC resource.
It uses noble to connect via BLE to the environmental sensor to get data.
There are temperature, humidity, air pressure and UV index sensor data that
the resource represents. To use, launch the nodejs application in the
homegateway with BLE capability. It has been tested on MinnowBoard Turbot &
Galileo with a CSR 4.0 BLE USB dongle. Edison is not yet supported.
* Modified README.md to describe how to use the sensor with the OCF server in the demo.
* Make the ocf server send simulated data.

Signed-off-by: Kien Dinh <kien.t.dinh@intel.com>
Signed-off-by: Theresa <theresa.shan@intel.com>