Set-up dedicated folder for the GW code and add relevant documentation (#62)
The folder structure was not very logical anymore and it was the result
of a variety of additional components that have been added over the past
few months. Some parts of the documentation (README.md) was out-of-date
or non-existent for some parts. This commit separates the various pieces
in a more logical manner and adds some (at times basic) documentation
via various README.md files located in the various folders.

With one minor exception, no changes to the actual code was made (not
needed). The only exception to that is the path used in a service file
and the main entry point for the GW server that has been renamed into
gateway-server.js

Signed-off-by: Geoffroy Van Cutsem <geoffroy.vancutsem@intel.com>